GamesPresenter - #2 Pull Request 리뷰 수정
