+++ Dailymlb/app/src/main/java/com/jmp/dailymlb/iface/APIService.java
@@ -5,7 +5,6 @@ import com.jmp.dailymlb.model.GameScore;
 
 import java.util.List;
 
-import okhttp3.ResponseBody;
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
