+++ Dailymlb/app/src/main/java/com/jmp/dailymlb/presenter/GamesContract.java
@@ -1,8 +1,6 @@
 package com.jmp.dailymlb.presenter;
 
 import com.jmp.dailymlb.model.GameScore;
-import com.jmp.dailymlb.model.source.gameScore.GameScoreRepository;
-
 import java.util.List;
 
 public interface GamesContract {
@@ -13,7 +11,6 @@ public interface GamesContract {
     interface Presenter {
         void attachView(View view);
         void detachView();
-        void setGameScoreData(GameScoreRepository gameScoreData);
         void loadGamesByDate();
 
     }
